requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Data
pandas>=2.0.0
wfdb>=4.1.0          # for loading PTB-XL real ECG data

# Visualisation
matplotlib>=3.7.0
Pillow>=10.0.0

# LLM
anthropic>=0.30.0

# UI
streamlit>=1.35.0

# Utilities
tqdm>=4.65.0
